qr-tag_overlay/app/requirements.txt
aiohttp==3.9.5
Pillow==10.4.0
pyzbar==0.1.9
